+++ RoslynMemberSorter/Enums/FieldMutability.cs
@@ -6,17 +6,17 @@ namespace RoslynMemberSorter.Enums;
 public enum FieldMutability
 {
 	/// <summary>
-	/// The field is declared with the const keyword.
+	/// The field is declared with the <see langword="const" /> keyword.
 	/// </summary>
 	Const,
 
 	/// <summary>
-	/// The field is declared with neither the const keyword nor the readonly keyword.
+	/// The field is declared with neither the <see langword="const" /> keyword nor the <see langword="readonly" /> keyword.
 	/// </summary>
 	Mutable,
 
 	/// <summary>
-	/// The field is declared with the readonly keyword.
+	/// The field is declared with the <see langword="readonly" /> keyword.
 	/// </summary>
 	ReadOnly
 }